+++ devices/MK28F15/drivers/fsl_dspi_edma.c
@@ -1,35 +1,9 @@
 /*
- * The Clear BSD License
  * Copyright (c) 2015, Freescale Semiconductor, Inc.
  * Copyright 2016-2017 NXP
  * All rights reserved.
  *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted (subject to the limitations in the disclaimer below) provided
- *  that the following conditions are met:
- *
- * o Redistributions of source code must retain the above copyright notice, this list
- *   of conditions and the following disclaimer.
- *
- * o Redistributions in binary form must reproduce the above copyright notice, this
- *   list of conditions and the following disclaimer in the documentation and/or
- *   other materials provided with the distribution.
- *
- * o Neither the name of the copyright holder nor the names of its
- *   contributors may be used to endorse or promote products derived from this
- *   software without specific prior written permission.
- *
- * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS LICENSE.
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * SPDX-License-Identifier: BSD-3-Clause
  */
 
 #include "fsl_dspi_edma.h"
@@ -94,6 +68,26 @@ static dspi_slave_edma_private_handle_t s_dspiSlaveEdmaPrivateHandle[FSL_FEATURE
 * Code
 ***********************************************************************************************************************/
 
+/*!
+ * brief Initializes the DSPI master eDMA handle.
+ *
+ * This function initializes the DSPI eDMA handle which can be used for other DSPI transactional APIs.  Usually, for a
+ * specified DSPI instance, call this API once to get the initialized handle.
+ *
+ * Note that DSPI eDMA has separated (RX and TX as two sources) or shared (RX  and TX are the same source) DMA request
+ * source.
+ * (1) For the separated DMA request source, enable and set the RX DMAMUX source for edmaRxRegToRxDataHandle and
+ * TX DMAMUX source for edmaIntermediaryToTxRegHandle.
+ * (2) For the shared DMA request source, enable and set the RX/RX DMAMUX source for the edmaRxRegToRxDataHandle.
+ *
+ * param base DSPI peripheral base address.
+ * param handle DSPI handle pointer to dspi_master_edma_handle_t.
+ * param callback DSPI callback.
+ * param userData A callback function parameter.
+ * param edmaRxRegToRxDataHandle edmaRxRegToRxDataHandle pointer to edma_handle_t.
+ * param edmaTxDataToIntermediaryHandle edmaTxDataToIntermediaryHandle pointer to edma_handle_t.
+ * param edmaIntermediaryToTxRegHandle edmaIntermediaryToTxRegHandle pointer to edma_handle_t.
+ */
 void DSPI_MasterTransferCreateHandleEDMA(SPI_Type *base,
                                          dspi_master_edma_handle_t *handle,
                                          dspi_master_edma_transfer_callback_t callback,
@@ -125,6 +119,17 @@ void DSPI_MasterTransferCreateHandleEDMA(SPI_Type *base,
     handle->edmaIntermediaryToTxRegHandle = edmaIntermediaryToTxRegHandle;
 }
 
+/*!
+ * brief DSPI master transfer data using eDMA.
+ *
+ * This function transfers data using eDMA. This is a non-blocking function, which returns right away. When all data
+ * is transferred, the callback function is called.
+ *
+ * param base DSPI peripheral base address.
+ * param handle A pointer to the dspi_master_edma_handle_t structure which stores the transfer state.
+ * param transfer A pointer to the dspi_transfer_t structure.
+ * return status of status_t.
+ */
 status_t DSPI_MasterTransferEDMA(SPI_Type *base, dspi_master_edma_handle_t *handle, dspi_transfer_t *transfer)
 {
     assert(handle);
@@ -169,7 +174,7 @@ status_t DSPI_MasterTransferEDMA(SPI_Type *base, dspi_master_edma_handle_t *hand
     dspi_command_data_config_t commandStruct;
     DSPI_StopTransfer(base);
     DSPI_FlushFifo(base, true, true);
-    DSPI_ClearStatusFlags(base, kDSPI_AllStatusFlag);
+    DSPI_ClearStatusFlags(base, (uint32_t)kDSPI_AllStatusFlag);
 
     commandStruct.whichPcs =
         (dspi_which_pcs_t)(1U << ((transfer->configFlags & DSPI_MASTER_PCS_MASK) >> DSPI_MASTER_PCS_SHIFT));
@@ -917,6 +922,17 @@ status_t DSPI_MasterTransferEDMA(SPI_Type *base, dspi_master_edma_handle_t *hand
     return kStatus_Success;
 }
 
+/*!
+ * brief Transfers a block of data using a eDMA method.
+ *
+ * This function transfers data using eDNA, the transfer mechanism is half-duplex. This is a non-blocking function,
+ * which returns right away. When all data is transferred, the callback function is called.
+ *
+ * param base DSPI base pointer
+ * param handle A pointer to the dspi_master_edma_handle_t structure which stores the transfer state.
+ * param transfer A pointer to the dspi_half_duplex_transfer_t structure.
+ * return status of status_t.
+ */
 status_t DSPI_MasterHalfDuplexTransferEDMA(SPI_Type *base,
                                            dspi_master_edma_handle_t *handle,
                                            dspi_half_duplex_transfer_t *xfer)
@@ -995,6 +1011,14 @@ static void EDMA_DspiMasterCallback(edma_handle_t *edmaHandle,
     }
 }
 
+/*!
+ * brief DSPI master aborts a transfer which is using eDMA.
+ *
+ * This function aborts a transfer which is using eDMA.
+ *
+ * param base DSPI peripheral base address.
+ * param handle A pointer to the dspi_master_edma_handle_t structure which stores the transfer state.
+ */
 void DSPI_MasterTransferAbortEDMA(SPI_Type *base, dspi_master_edma_handle_t *handle)
 {
     assert(handle);
@@ -1010,6 +1034,16 @@ void DSPI_MasterTransferAbortEDMA(SPI_Type *base, dspi_master_edma_handle_t *han
     handle->state = kDSPI_Idle;
 }
 
+/*!
+ * brief Gets the master eDMA transfer count.
+ *
+ * This function gets the master eDMA transfer count.
+ *
+ * param base DSPI peripheral base address.
+ * param handle A pointer to the dspi_master_edma_handle_t structure which stores the transfer state.
+ * param count A number of bytes transferred by the non-blocking transaction.
+ * return status of status_t.
+ */
 status_t DSPI_MasterTransferGetCountEDMA(SPI_Type *base, dspi_master_edma_handle_t *handle, size_t *count)
 {
     assert(handle);
@@ -1036,6 +1070,25 @@ status_t DSPI_MasterTransferGetCountEDMA(SPI_Type *base, dspi_master_edma_handle
     return kStatus_Success;
 }
 
+/*!
+ * brief Initializes the DSPI slave eDMA handle.
+ *
+ * This function initializes the DSPI eDMA handle which can be used for other DSPI transactional APIs.  Usually, for a
+ * specified DSPI instance, call this API once to get the initialized handle.
+ *
+ * Note that DSPI eDMA has separated (RN and TX in 2 sources) or shared (RX  and TX are the same source) DMA request
+ * source.
+ * (1)For the separated DMA request source, enable and set the RX DMAMUX source for edmaRxRegToRxDataHandle and
+ * TX DMAMUX source for edmaTxDataToTxRegHandle.
+ * (2)For the shared DMA request source,  enable and set the RX/RX DMAMUX source for the edmaRxRegToRxDataHandle.
+ *
+ * param base DSPI peripheral base address.
+ * param handle DSPI handle pointer to dspi_slave_edma_handle_t.
+ * param callback DSPI callback.
+ * param userData A callback function parameter.
+ * param edmaRxRegToRxDataHandle edmaRxRegToRxDataHandle pointer to edma_handle_t.
+ * param edmaTxDataToTxRegHandle edmaTxDataToTxRegHandle pointer to edma_handle_t.
+ */
 void DSPI_SlaveTransferCreateHandleEDMA(SPI_Type *base,
                                         dspi_slave_edma_handle_t *handle,
                                         dspi_slave_edma_transfer_callback_t callback,
@@ -1062,6 +1115,19 @@ void DSPI_SlaveTransferCreateHandleEDMA(SPI_Type *base,
     handle->edmaTxDataToTxRegHandle = edmaTxDataToTxRegHandle;
 }
 
+/*!
+ * brief DSPI slave transfer data using eDMA.
+ *
+ * This function transfers data using eDMA. This is a non-blocking function, which returns right away. When all data
+ * is transferred, the callback function is called.
+ * Note that the slave eDMA transfer doesn't support transfer_size is 1 when the bitsPerFrame is greater
+ * than eight.
+
+ * param base DSPI peripheral base address.
+ * param handle A pointer to the dspi_slave_edma_handle_t structure which stores the transfer state.
+ * param transfer A pointer to the dspi_transfer_t structure.
+ * return status of status_t.
+ */
 status_t DSPI_SlaveTransferEDMA(SPI_Type *base, dspi_slave_edma_handle_t *handle, dspi_transfer_t *transfer)
 {
     assert(handle);
@@ -1152,7 +1218,7 @@ status_t DSPI_SlaveTransferEDMA(SPI_Type *base, dspi_slave_edma_handle_t *handle
     DSPI_StopTransfer(base);
 
     DSPI_FlushFifo(base, true, true);
-    DSPI_ClearStatusFlags(base, kDSPI_AllStatusFlag);
+    DSPI_ClearStatusFlags(base, (uint32_t)kDSPI_AllStatusFlag);
 
     DSPI_DisableDMA(base, kDSPI_RxDmaEnable | kDSPI_TxDmaEnable);
 
@@ -1405,6 +1471,14 @@ static void EDMA_DspiSlaveCallback(edma_handle_t *edmaHandle,
     }
 }
 
+/*!
+ * brief DSPI slave aborts a transfer which is using eDMA.
+ *
+ * This function aborts a transfer which is using eDMA.
+ *
+ * param base DSPI peripheral base address.
+ * param handle A pointer to the dspi_slave_edma_handle_t structure which stores the transfer state.
+ */
 void DSPI_SlaveTransferAbortEDMA(SPI_Type *base, dspi_slave_edma_handle_t *handle)
 {
     assert(handle);
@@ -1419,6 +1493,16 @@ void DSPI_SlaveTransferAbortEDMA(SPI_Type *base, dspi_slave_edma_handle_t *handl
     handle->state = kDSPI_Idle;
 }
 
+/*!
+ * brief Gets the slave eDMA transfer count.
+ *
+ * This function gets the slave eDMA transfer count.
+ *
+ * param base DSPI peripheral base address.
+ * param handle A pointer to the dspi_slave_edma_handle_t structure which stores the transfer state.
+ * param count A number of bytes transferred so far by the non-blocking transaction.
+ * return status of status_t.
+ */
 status_t DSPI_SlaveTransferGetCountEDMA(SPI_Type *base, dspi_slave_edma_handle_t *handle, size_t *count)
 {
     assert(handle);
